NEW : Change code to a "Controller-Service" architectural, controller handle incoming requests and service performs operations on the database.
New : hierarchy of objects (OOP) for service

Update : file name Camel case to snake case
