+++ services/absent_service.py
@@ -1,0 +1,10 @@
+# from services.main_service import Service
+#
+# import connect_pg
+# import psycopg2
+# import requests
+# import hashlib
+# import json
+#
+# class absent_service(Service):
+#
